+++ CHANGELOG.md
@@ -1,3 +1,9 @@
+## poetry-dynamic-versioning-plugin: Unreleased
+
+* Fixed:
+  * `poetry` did not work correctly when in a folder without a pyproject.toml.
+  * `poetry plugin show` did not work correctly.
+
 ## poetry-dynamic-versioning: v0.17.1 & poetry-dynamic-versioning-plugin: v0.3.1
 
 * Fixed: